// src/types/shared.d.ts
export type numberInString = string | number;
export type ExchangeSymbol = string;
export type BooleanString = 'true' | 'false';
export type BooleanStringCapitalised = 'TRUE' | 'FALSE';
export type BinanceBaseUrlKey = 'spot' | 'spot1' | 'spot2' | 'spot3' | 'spot4' | 'usdmtest' | 'usdm' | 'coinm' | 'coinmtest' | 'voptions' | 'voptionstest' | 'papi';
export type OrderTimeInForce = 'GTC' | 'IOC' | 'FOK' | 'GTX' | 'GTE_GTC' | 'GTD';
export type StringBoolean = 'TRUE' | 'FALSE';
export type SideEffects = 'MARGIN_BUY' | 'AUTO_REPAY' | 'NO_SIDE_EFFECT';
export type OrderResponseType = 'ACK' | 'RESULT' | 'FULL';
export type OrderIdProperty = 'newClientOrderId' | 'newClientStrategyId' | 'listClientOrderId' | 'limitClientOrderId' | 'stopClientOrderId' | 'aboveClientOrderId' | 'belowClientOrderId' | 'workingClientOrderId' | 'pendingAboveClientOrderId' | 'pendingBelowClientOrderId' | 'pendingClientOrderId' | 'clientAlgoId';
export type OrderSide = 'BUY' | 'SELL';
// ...
export interface BasicSymbolParam {
    symbol: string;
    isIsolated?: StringBoolean;
}
export interface SymbolArrayParam {
    symbols: string[];
}
export interface BasicAssetPaginatedParams {
    asset?: string;
    startTime?: number;
    endTime?: number;
    limit?: number;
}
export interface BasicSymbolPaginatedParams {
    symbol?: string;
    startTime?: number;
    endTime?: number;
    limit?: number;
}
export interface SymbolPrice {
    symbol: string;
    price: numberInString;
}
export interface OrderBookParams {
    symbol: string;
    limit?: 5 | 10 | 20 | 50 | 100 | 500 | 1000 | 5000;
}
export type KlineInterval = '1s' | '1m' | '3m' | '5m' | '15m' | '30m' | '1h' | '2h' | '4h' | '6h' | '8h' | '12h' | '1d' | '3d' | '1w' | '1M';
export interface GetOrderParams {
    symbol: string;
    orderId?: number;
    origClientOrderId?: string;
    isIsolated?: StringBoolean;
}
export interface GetOrderModifyHistoryParams {
    symbol: string;
    orderId?: number;
    origClientOrderId?: string;
    isIsolated?: StringBoolean;
    startTime?: number;
    endTime?: number;
    limit?: number;
}
export interface HistoricalTradesParams {
    symbol: string;
    limit?: number;
    fromId?: number;
}
export interface KlinesParams {
    symbol: string;
    interval: KlineInterval;
    startTime?: number;
    endTime?: number;
    limit?: number;
}
export type Kline = [
    number,
    numberInString,
    numberInString,
    numberInString,
    numberInString,
    numberInString,
    number,
    numberInString,
    number,
    numberInString,
    numberInString,
    numberInString
];
export type FuturesKline = Kline;
export interface RecentTradesParams {
    symbol: string;
    limit?: number;
}
export interface CancelOrderParams {
    symbol: string;
    orderId?: number;
    origClientOrderId?: string;
    newClientOrderId?: string;
    isIsolated?: StringBoolean;
}
export interface CancelOCOParams {
    symbol: string;
    isIsolated?: string;
    orderListId?: number;
    listClientOrderId?: string;
    newClientOrderId?: string;
}
export interface NewOCOParams {
    symbol: string;
    listClientOrderId?: string;
    side: OrderSide;
    quantity: number;
    limitClientOrderId?: string;
    limitStrategyId?: number;
    limitStrategyType?: number;
    price: number;
    limitIcebergQty?: number;
    trailingDelta?: number;
    stopClientOrderId?: string;
    stopPrice: number;
    stopStrategyId?: number;
    stopStrategyType?: number;
    stopLimitPrice?: number;
    stopIcebergQty?: number;
    stopLimitTimeInForce?: OrderTimeInForce;
    newOrderRespType?: OrderResponseType;
    isIsolated?: StringBoolean;
    sideEffectType?: SideEffects;
}
// ...
export interface SymbolFromPaginatedRequestFromId {
    symbol: string;
    fromId?: number;
    startTime?: number;
    endTime?: number;
    limit?: number;
}