// src/types/futures.d.ts
export type FuturesContractType = 'PERPETUAL' | 'CURRENT_MONTH' | 'NEXT_MONTH' | 'CURRENT_QUARTER' | 'NEXT_QUARTER';
export interface ContinuousContractKlinesParams {
    pair: string;
    contractType: FuturesContractType;
    interval: KlineInterval;
    startTime?: number;
    endTime?: number;
    limit?: number;
}
export interface IndexPriceKlinesParams {
    pair: string;
    interval: KlineInterval;
    startTime?: number;
    endTime?: number;
    limit?: number;
}
export interface SymbolKlinePaginatedParams {
    symbol: string;
    interval: KlineInterval;
    startTime?: number;
    endTime?: number;
    limit?: number;
}
export interface FuturesDataPaginatedParams {
    symbol: string;
    period: '5m' | '15m' | '30m' | '1h' | '2h' | '4h' | '6h' | '12h' | '1d';
    limit?: number;
    startTime?: number;
    endTime?: number;
}
export interface FuturesCoinMTakerBuySellVolumeParams {
    pair: string;
    contractType: 'ALL' | 'CURRENT_QUARTER' | 'NEXT_QUARTER' | 'PERPETUAL';
    period: '5m' | '15m' | '30m' | '1h' | '2h' | '4h' | '6h' | '12h' | '1d';
    limit?: number;
    startTime?: number;
    endTime?: number;
}
export interface FuturesCoinMBasisParams {
    pair: string;
    contractType: 'CURRENT_QUARTER' | 'NEXT_QUARTER' | 'PERPETUAL';
    period: '5m' | '15m' | '30m' | '1h' | '2h' | '4h' | '6h' | '12h' | '1d';
    limit?: number;
    startTime?: number;
    endTime?: number;
}
export declare enum EnumDualSideMode {
    HedgeMode = "true",
    OneWayMode = "false"
}
export type DualSideMode = `${EnumDualSideMode}`;
export declare enum EnumMultiAssetMode {
    MultiAssetsMode = "true",
    SingleAssetsMode = "false"
}
export type MultiAssetsMode = `${EnumMultiAssetMode}`;
export type PositionSide = 'BOTH' | 'LONG' | 'SHORT';
export type MarginType = 'ISOLATED' | 'CROSSED';
export type WorkingType = 'MARK_PRICE' | 'CONTRACT_PRICE';
export type FuturesOrderType = 'LIMIT' | 'MARKET' | 'STOP' | 'STOP_MARKET' | 'TAKE_PROFIT' | 'TAKE_PROFIT_MARKET' | 'TRAILING_STOP_MARKET';
export type PriceMatchMode = 'NONE' | 'OPPONENT' | 'OPPONENT_5' | 'OPPONENT_10' | 'OPPONENT_20' | 'QUEUE' | 'QUEUE_5' | 'QUEUE_10' | 'QUEUE_20';
export interface NewFuturesOrderParams<numberType = number> {
    symbol: string;
    side: OrderSide;
    positionSide?: PositionSide;
    type: FuturesOrderType;
    timeInForce?: OrderTimeInForce;
    quantity?: numberType;
    reduceOnly?: BooleanString;
    price?: numberType;
    newClientOrderId?: string;
    stopPrice?: numberType;
    closePosition?: BooleanString;
    activationPrice?: numberType;
    callbackRate?: numberType;
    workingType?: WorkingType;
    priceProtect?: BooleanStringCapitalised;
    newOrderRespType?: OrderResponseType;
    selfTradePreventionMode?: SelfTradePreventionMode;
    priceMatch?: PriceMatchMode;
}
// ...
export type ContactType = 'PERPETUAL' | 'CURRENT_MONTH' | 'NEXT_MONTH' | 'CURRENT_QUARTER' | 'NEXT_QUARTER';
export type ContractStatus = 'PENDING_TRADING' | 'TRADING' | 'PRE_DELIVERING' | 'DELIVERING' | 'DELIVERED' | 'CANCELLED' | 'PRE_SETTLE' | 'SETTLING' | 'CLOSE';
// ...
export type FuturesSymbolFilter = SymbolPriceFilter | FuturesSymbolPercentPriceFilter | SymbolLotSizeFilter | FuturesSymbolMinNotionalFilter | SymbolIcebergPartsFilter | SymbolMarketLotSizeFilter | FuturesSymbolMaxOrdersFilter | FuturesSymbolMaxAlgoOrdersFilter | SymbolMaxIcebergOrdersFilter | SymbolMaxPositionFilter;
export interface FuturesSymbolExchangeInfo {
    symbol: string;
    pair: string;
    contractType: ContactType;
    deliveryDate: number;
    onboardDate: number;
    status: ContractStatus;
    maintMarginPercent: numberInString;
    requiredMarginPercent: numberInString;
    baseAsset: string;
    quoteAsset: string;
    marginAsset: string;
    pricePrecision: number;
    quantityPrecision: number;
    baseAssetPrecision: number;
    quotePrecision: number;
    underlyingType: 'COIN' | 'INDEX';
    underlyingSubType: string[];
    settlePlan: number;
    triggerProtect: numberInString;
    filters: FuturesSymbolFilter[];
    OrderType: OrderType[];
    timeInForce: OrderTimeInForce[];
    liquidationFee: numberInString;
    marketTakeBound: numberInString;
    contractSize?: number;
}
export interface FuturesExchangeInfo {
    exchangeFilters: ExchangeFilter[];
    rateLimits: RateLimiter[];
    serverTime: number;
    assets: any[];
    symbols: FuturesSymbolExchangeInfo[];
    timezone: string;
}
export interface FuturesOrderBook {
    lastUpdateId: number;
    E: number;
    T: number;
    bids: OrderBookRow[];
    asks: OrderBookRow[];
}
export interface RawFuturesTrade {
    id: number;
    price: numberInString;
    qty: numberInString;
    quoteQty: numberInString;
    time: number;
    isBuyerMaker: boolean;
}
export interface AggregateFuturesTrade {
    a: number;
    p: numberInString;
    q: numberInString;
    f: number;
    l: number;
    T: number;
    M: boolean;
}
export interface MarkPrice {
    symbol: string;
    markPrice: numberInString;
    indexPrice: numberInString;
    estimatedSettlePrice: numberInString;
    lastFundingRate: numberInString;
    interestRate: numberInString;
    nextFundingTime: number;
    time: number;
}
export interface FundingRateHistory {
    symbol: string;
    fundingRate: numberInString;
    fundingTime: number;
}
// ...
export interface ChangeStats24hr {
    symbol: string;
    priceChange: numberInString;
    priceChangePercent: numberInString;
    weightedAvgPrice: numberInString;
    lastPrice: numberInString;
    lastQty: numberInString;
    openPrice: numberInString;
    highPrice: numberInString;
    lowPrice: numberInString;
    volume: numberInString;
    quoteVolume: numberInString;
    openTime: number;
    closeTime: number;
    firstId: number;
    lastId: number;
    count: number;
}
// ...
export interface QuarterlyContractSettlementPrice {
    deliveryTime: number;
    deliveryPrice: string;
}
// ...
export interface IndexPriceConstituent {
    exchange: string;
    symbol: string;
}
export interface IndexPriceConstituents {
    symbol: string;
    time: number;
    constituents: IndexPriceConstituent[];
}